Revert "Merge tag 'LA.UM.7.3.r1-05100-sdm845.0' of https://source.codeaurora.org/quic/la/kernel/msm-4.9"
This reverts commit 84dd14c82f9c780a22479d58d25ec3acf9ca0626, reversing
changes made to 684116bef851f5b450a14738556018e2a7be52e6.

+++ arch/x86/events/perf_event.h
@@ -548,7 +548,7 @@ struct x86_pmu {
 	struct x86_pmu_quirk *quirks;
 	int		perfctr_second_write;
 	bool		late_ack;
-	u64		(*limit_period)(struct perf_event *event, u64 l);
+	unsigned	(*limit_period)(struct perf_event *event, unsigned l);
 
 	/*
 	 * sysfs attrs
